pyproject.toml
[tool.poetry]
name = "rua-bbps-experiments"
version = "0.1.0"
description = "RUA-BBPS experiment code with Hydra and WandB"
authors = ["Researcher <researcher@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
torch = "^1.12"
torchvision = "^0.13"
hydra-core = "^1.3"
wandb = "^0.15"
transformers = "^4.30"
optuna = "^3.0"
openai = "^0.27"
matplotlib = "^3.7"
seaborn = "^0.12"
pandas = "^2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
